python/requirements.txt
openai>=1.3.0
pymongo>=4.6.0
pinecone>=7.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=11.0.0
pytesseract>=0.3.0
python-pptx>=0.6.0
opencv-python>=4.8.0
SpeechRecognition>=3.10.0
moviepy>=1.0.0
librosa>=0.10.0
soundfile>=0.12.0
scikit-learn>=1.3.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
pdfplumber>=0.9.0